# docs/specs/001-sakuraplayer-v1/contracts/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: SakuraPlayer API
  version: 1.2.0
  description: >-
    Private single-administrator API for the SakuraPlayer v1 backend. Video bytes
    are never proxied by this API; signed stream endpoints return 302 redirects.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /movies/{movie_id}/play-requests:
    post:
      tags: [Cache]
      operationId: createPlayRequest
      description: >-
        The only operation that may create a 115 offline job. The source_id must
        belong to the movie and be an active AVdb source. A started disposition
        allows at most 60 seconds of client blocking; the backend job continues
        after that client wait expires. Replaying the same Idempotency-Key with
        the same movie/source returns the original job in any state; reusing the
        key for another payload returns idempotency_conflict.
      parameters:
        - $ref: '#/components/parameters/MovieId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [source_id]
              properties:
                source_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Existing ready or active task reused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayRequestResult'
        '202':
          description: New started or queued cache job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayRequestResult'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /cache-jobs:
    get:
      tags: [Cache]
      operationId: listCacheJobs
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CacheJobStatus'
          style: form
          explode: false
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Task snapshot and fixed capacity counters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheJobPage'
  /cache-jobs/{cache_job_id}:
    get:
      tags: [Cache]
      operationId: getCacheJob
      parameters:
        - $ref: '#/components/parameters/CacheJobId'
      responses:
        '200':
          description: Full safe task snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheJob'
        '404':
          $ref: '#/components/responses/NotFound'
  /cache-jobs/{cache_job_id}/cancel:
    post:
      tags: [Cache]
      operationId: cancelCacheJob
      parameters:
        - $ref: '#/components/parameters/CacheJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmed]
              properties:
                confirmed:
                  type: boolean
                  const: true
      responses:
        '202':
          description: Cancellation and safe cleanup accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheJob'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
  /cache-jobs/{cache_job_id}/cleanup:
    post:
      tags: [Cache]
      operationId: cleanupCacheJob
      description: >-
        Requests the same proof-based cleanup used by maintenance. Ready and
        awaiting-selection jobs enter cleaning; cleanup_failed retries create a
        new fenced attempt. Active playback leases reject the request.
      parameters:
        - $ref: '#/components/parameters/CacheJobId'
      responses:
        '202':
          description: Safe cleanup accepted when no active lease exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheJob'
        '409':
          $ref: '#/components/responses/Conflict'
  /cache-jobs/{cache_job_id}/media-selection:
    put:
      tags: [Cache]
      operationId: selectCacheMedia
      parameters:
        - $ref: '#/components/parameters/CacheJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [media_ids]
              properties:
                media_ids:
                  type: array
                  minItems: 1
                  maxItems: 100
                  uniqueItems: true
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Selected media or ordered segment queue saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheJob'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: >-
        Opaque versioned keyset cursor bound to the endpoint's normalized query,
        filters and sort. Malformed or cross-query cursors return validation_failed.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 24
    MovieId:
      name: movie_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CacheJobId:
      name: cache_job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._~-]{16,128}$'
...
  responses:
...
    NotFound:
      description: Requested resource does not exist or is not visible.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: State, capacity, idempotency or ownership conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationError:
      description: Request cannot be validated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          pattern: '^[a-z0-9_]+$'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
...
    CacheJobStatus:
      type: string
      enum:
        - queued
        - submitting
        - offlining
        - submit_uncertain
        - resolving
        - awaiting_selection
        - ready
        - cancelling
        - cleaning
        - cleanup_failed
        - failed
        - cleaned
        - detached
    RemoteMedia:
      type: object
      required: [id, candidate_id, name, size_bytes, sequence_no, is_valid]
      properties:
        id:
          type: string
          format: uuid
        candidate_id:
          type: string
          format: uuid
        name:
          type: string
        size_bytes:
          type: integer
          format: int64
        duration_seconds:
          type: [integer, 'null']
          format: int64
        sequence_no:
          type: integer
        is_valid:
          type: boolean
    SubtitleOption:
      type: object
      required: [id, media_id, name, format, selected_by_default]
      properties:
        id:
          type: string
          format: uuid
        media_id:
          type: [string, 'null']
          format: uuid
          description: Null means the subtitle is available to every media session in this manifest.
        name:
          type: string
        format:
          type: string
          enum: [srt, ass, ssa, vtt]
        language:
          type: [string, 'null']
        selected_by_default:
          type: boolean
    CacheJob:
      type: object
      required:
        [id, movie_id, source_id, status, remote_percent, media_candidates, selected_media_ids, subtitles, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        movie_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/CacheJobStatus'
        remote_percent:
          type: number
          minimum: 0
          maximum: 100
        error_code:
          type: [string, 'null']
        media_candidates:
          type: array
          items:
            $ref: '#/components/schemas/RemoteMedia'
        selected_media_ids:
          type: array
          items:
            type: string
            format: uuid
        subtitles:
          type: array
          items:
            $ref: '#/components/schemas/SubtitleOption'
        ready_at:
          type: [string, 'null']
          format: date-time
        expires_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CacheCapacity:
      type: object
      required: [running, running_limit, queued, queued_limit, ready, ready_limit]
      properties:
        running:
          type: integer
        running_limit:
          type: integer
          const: 2
        queued:
          type: integer
        queued_limit:
          type: integer
          const: 10
        ready:
          type: integer
        ready_limit:
          type: integer
          default: 20
    CacheJobPage:
      type: object
      required: [items, capacity]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CacheJob'
        capacity:
          $ref: '#/components/schemas/CacheCapacity'
        next_cursor:
          type: [string, 'null']
    PlayRequestResult:
      type: object
      required: [disposition, cache_job]
      properties:
        disposition:
          type: string
          enum: [ready, started, queued, reused]
        wait_deadline:
          type: [string, 'null']
          format: date-time
          description: Present only for a newly created started disposition; exactly 60 seconds from the server response clock and never persisted.
        cache_job:
          $ref: '#/components/schemas/CacheJob'
...